# skillmeat/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillMeat API",
    "description": "REST API for SkillMeat collection manager",
    "version": "0.1.0-alpha"
  },
  "paths": {
...
    "/api/v1/analytics/summary": {
      "get": {
        "tags": [
          "analytics",
          "analytics"
        ],
        "summary": "Get analytics summary",
        "description": "Retrieve overall statistics and analytics summary",
        "operationId": "get_analytics_summary_api_v1_analytics_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved analytics summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsSummaryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Analytics disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/top-artifacts": {
      "get": {
        "tags": [
          "analytics",
          "analytics"
        ],
        "summary": "Get top artifacts by usage",
        "description": "Retrieve most used artifacts sorted by usage frequency",
        "operationId": "get_top_artifacts_api_v1_analytics_top_artifacts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page (max 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items per page (max 100)"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination (next page)",
              "title": "After"
            },
            "description": "Cursor for pagination (next page)"
          },
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by artifact type (skill, command, agent)",
              "title": "Artifact Type"
            },
            "description": "Filter by artifact type (skill, command, agent)"
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved top artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopArtifactsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Analytics disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/trends": {
      "get": {
        "tags": [
          "analytics",
          "analytics"
        ],
        "summary": "Get usage trends over time",
        "description": "Retrieve time-series usage data for trend analysis",
        "operationId": "get_usage_trends_api_v1_analytics_trends_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Aggregation period (hour, day, week, month)",
              "default": "day",
              "title": "Period"
            },
            "description": "Aggregation period (hour, day, week, month)"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "Number of days of history (max 365)",
              "default": 30,
              "title": "Days"
            },
            "description": "Number of days of history (max 365)"
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved usage trends",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Analytics disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyticsSummaryResponse": {
        "properties": {
          "total_collections": {
            "type": "integer",
            "title": "Total Collections",
            "description": "Total number of collections",
            "examples": [
              1
            ]
          },
          "total_artifacts": {
            "type": "integer",
            "title": "Total Artifacts",
            "description": "Total number of artifacts across all collections",
            "examples": [
              15
            ]
          },
          "total_deployments": {
            "type": "integer",
            "title": "Total Deployments",
            "description": "Total number of deployments",
            "examples": [
              42
            ]
          },
          "total_events": {
            "type": "integer",
            "title": "Total Events",
            "description": "Total number of tracked events",
            "examples": [
              1337
            ]
          },
          "artifacts_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Artifacts By Type",
            "description": "Count of artifacts by type",
            "examples": [
              {
                "agent": 2,
                "command": 3,
                "skill": 10
              }
            ]
          },
          "recent_activity_count": {
            "type": "integer",
            "title": "Recent Activity Count",
            "description": "Number of events in the last 24 hours",
            "examples": [
              23
            ]
          },
          "most_deployed_artifact": {
            "type": "string",
            "title": "Most Deployed Artifact",
            "description": "Name of the most frequently deployed artifact",
            "examples": [
              "pdf-skill"
            ]
          },
          "last_activity": {
            "type": "string",
            "format": "date-time",
            "title": "Last Activity",
            "description": "Timestamp of most recent activity"
          }
        },
        "type": "object",
        "required": [
          "total_collections",
          "total_artifacts",
          "total_deployments",
          "total_events",
          "artifacts_by_type",
          "recent_activity_count",
          "most_deployed_artifact",
          "last_activity"
        ],
        "title": "AnalyticsSummaryResponse",
        "description": "Overall analytics summary.\n\nProvides high-level statistics about the collection and usage patterns.",
        "example": {
          "artifacts_by_type": {
            "agent": 2,
            "command": 3,
            "skill": 10
          },
          "last_activity": "2024-11-16T15:30:00Z",
          "most_deployed_artifact": "pdf-skill",
          "recent_activity_count": 23,
          "total_artifacts": 15,
          "total_collections": 1,
          "total_deployments": 42,
          "total_events": 1337
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type or code",
            "examples": [
              "NOT_FOUND"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message",
            "examples": [
              "Collection not found"
            ]
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional error details (optional)",
            "examples": [
              {
                "collection_id": "abc123"
              }
            ]
          },
          "validation_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Errors",
            "description": "Validation errors for invalid requests"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response envelope.\n\nAll API errors return this structure for consistent error handling\non the client side.",
        "example": {
          "error": "VALIDATION_ERROR",
          "message": "Invalid request parameters",
          "validation_errors": [
            {
              "field": "name",
              "message": "Field is required",
              "type": "value_error.missing"
            }
          ]
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TopArtifactsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TopArtifactItem"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items for this page"
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination metadata"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page_info"
        ],
        "title": "TopArtifactsResponse",
        "description": "Paginated response for top artifacts by usage.\n\nReturns artifacts sorted by usage frequency with pagination support."
      },
...
      "TrendsResponse": {
        "properties": {
          "period_type": {
            "type": "string",
            "title": "Period Type",
            "description": "Aggregation period (hour, day, week, month)",
            "examples": [
              "day"
            ]
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date",
            "description": "Start of trend period"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date",
            "description": "End of trend period"
          },
          "data_points": {
            "items": {
              "$ref": "#/components/schemas/TrendDataPoint"
            },
            "type": "array",
            "title": "Data Points",
            "description": "Time-series data points"
          },
          "total_periods": {
            "type": "integer",
            "title": "Total Periods",
            "description": "Number of periods in the response",
            "examples": [
              30
            ]
          }
        },
        "type": "object",
        "required": [
          "period_type",
          "start_date",
          "end_date",
          "data_points",
          "total_periods"
        ],
        "title": "TrendsResponse",
        "description": "Response for usage trends over time.\n\nProvides time-series data for usage patterns and trends.",
        "example": {
          "data_points": [
            {
              "deployment_count": 5,
              "period": "day",
              "timestamp": "2024-11-16T00:00:00Z",
              "top_artifact": "pdf-skill",
              "unique_artifacts": 3,
              "usage_count": 42
            }
          ],
          "end_date": "2024-11-16T00:00:00Z",
          "period_type": "day",
          "start_date": "2024-10-17T00:00:00Z",
          "total_periods": 30
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}